Added save and loading functionality. Refactored UserInput class into seperate classes and placed in package
