+++ src/main/java/Main.java
@@ -1,4 +1,4 @@
-
+import UserInput.UserInput;
 
 public class Main {
